Move api validation into the api
