+++ api/v1alpha1/provisioning_webhook_test.go
@@ -50,6 +50,7 @@ func TestProvisioningValidateCreate(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			tt.p.Spec = *disabledProvisioning().build()
 			if err := tt.p.ValidateCreate(); !errorContains(err, tt.wantErr) {
 				t.Errorf("Provisioning.ValidateCreate() error = %v, wantErr %v", err, tt.wantErr)
 			}
